keyMera/src/main/java/org/iotmit/service/BoardService.java
package org.iotmit.service;

import java.util.List;

import org.iotmit.domain.BoardVO;
import org.iotmit.domain.Criteria;

public interface BoardService {
	//1.목록보기
	public List<BoardVO> getList(Criteria cri );
	//2.게시글등록
	public void register(BoardVO board);
	//3.게시글보기
	public BoardVO get(long bno);
	//4.게시글수정
	public boolean modify(BoardVO board);
	//5.게시글삭제
	public boolean remove(long bno);
	//6.전체 게시판 글갰수
	public long scount();
	
	public long admin_scount();
	
	public long search_count(String writter);
	
	//public BoardVO get(int num1,int num2);
		
}
